Split base and derived class implementations. [#CLICKHOUSE-2902]
