+++ dbms/src/Storages/MergeTree/MergeTreeReadPool.h
@@ -68,7 +68,7 @@ public:
         const std::size_t threads, const std::size_t sum_marks, const std::size_t min_marks_for_concurrent_read,
         RangesInDataParts parts, MergeTreeData & data, const ExpressionActionsPtr & prewhere_actions,
         const String & prewhere_column_name, const bool check_columns, const Names & column_names,
-        const BackoffSettings & backoff_settings,
+        const BackoffSettings & backoff_settings, size_t preferred_block_size_bytes,
         const bool do_not_steal_tasks = false);
 
     MergeTreeReadTaskPtr getTask(const std::size_t min_marks_to_read, const std::size_t thread);
@@ -92,12 +92,14 @@ private:
     MergeTreeData & data;
     Names column_names;
     bool do_not_steal_tasks;
+    bool predict_block_size_bytes;
     std::vector<NameSet> per_part_column_name_set;
     std::vector<NamesAndTypesList> per_part_columns;
     std::vector<NamesAndTypesList> per_part_pre_columns;
     /// @todo actually all of these values are either true or false for the whole query, thus no vector required
     std::vector<char> per_part_remove_prewhere_column;
     std::vector<char> per_part_should_reorder;
+    std::vector<MergeTreeBlockSizePredictorPtr> per_part_size_predictor;
 
     struct Part
     {
